Prepare 3.0.0 release and support for DC Platform 7 (#45)
Bitbucket 9 switched to Platform 7 and the plugin did no longer work. Getting this fixed was quite involved because our plugin was so dated, and Atlassian made a ton of changes to the platform in the last few years that we had to catch up to now.

Test plan: Plugin runs in the local bitbucket instance, our admin UI renders correctly, data persistence works, and I was able to correctly receive webhooks.